refactor(daemon): run library rescan in worker thread
when the daemon is told to run a library rescan, instead of running it and awaiting the result, we now run the task in a thread and don't wait for it to finish. This is important because a full rescan can easily take so long that tarpc would otherwise terminate the connection.
To prevent concurrent rescans, we use a lock that allows only one rescan to occur at any given moment.