feat(tui): add random view and integrate into sidebar and state management
Whenever the view is set as the active content view, it queries the daemon for a random album, song, or artist, the results are stored in the component props.

The view displays 3 options to the user, "Random Album", "Random Artist", and "Random Song", that when clicked navigate to the view for the album/artist/song that it got as described above.

The view gets new random items every time it is set as the active content view, so if you navigate away and go back you'll have different items than before.

fixes #200