refactor(udp): optimize buffer management in Sender and update publisher to use RwLock
give sender a buffer pool to allow `send` to use an immutable reference, enabling the Daemon to wrap its publisher with a RwLock instead of a Mutex. According to the benchmarks from the last commit this will be a huge performance win for the event notification system, especially going forward when more events are sent over the wire.