refactor(core): overhaul duration tracking
overhaul duration tracking,
removing the duration watcher thread entirely,
as well as the `duration_info` field previously used to track the
duration of the current song and amount of the song that's been played.

the `current_duration` field of the `DurationInfo` was only ever read by
the `state` command, which has access to the `current_song` field anyway
so we can just use that.

the `time_played` field of the `DurationInfo` was replaced with calls to
`player.get_pos()`, which was introduced in rodio v0.19.0.

One downside of this approach is that now we have a slight gap in
playback since we only start the next song when the current song is
completely done (as opposed to starting 100ms before it ends, like we
used to)