feat(daemon): integrate queue state persistence functionality
on startup, attempt to restore the previously save state, if any.
on exit, save the current state to the same file for next time.

Note: There is currently no mechanism to ensure that the songs being saved/loaded actually exist in the database, this was to keep this mechanism decoupled from the storage layer, the storage layer  itself should have enough error handling to handle things gracefully if, say, between the song being put in the state file and it being loaded again, it is deleted from the database

Note 2: if you do find yourself in this situation and it does cause a crash, open an issue and use the `--no-persistance` flag when starting the daemon and/or delete the mecomp.state.json file from the data directory