feat(tui): default to "connecting with retry" behavior when initializing client
when the TUI is started, it attempts to connect to the deamon,
previously it would start the daemon in the background if this failed,
but now it will just retry every second up to 5 times.
By enabling the `autostart-daemon` compile flag for `mecomp-tui`, the
old behavior can be restored.

Essentially, this makes the TUI behave more like the MPRIS layer when it
comes to connecting to the daemon.

next step is to refactor the MPRIS layer to use the new functions